requirements.txt
pycryptodome==3.19.0
Requests==2.31.0
